Add files via upload
This is a python data analyses program which utilizes the pandas library along with a Kaggle dataset of every NBA Finals series starting from 1980 - 2018. We would like to see how big of an impact winning Game 1 of an NBA Final could have on the probability of winning an NBA Championship. This project will not only analyze NBA Championship team data, but also the data for NBA Final runner ups